+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [0.1.5] - 2025-11-18
+
+### Added
+
+-   Add optional setting to use extended predicates and roll options
+
+### New and updated predicates
+
+-   Added predicates for PF2e playtest feats
+-   Added optional predicates for best lore, known spells, spellcasting traditions and proficiency, class HP, languages and senses.
+
 ## [0.1.4] - 2025-11-17
 
 ### Changes